Migrate pkg/networkservice/test -> pkg/networkservice/utils
Also testerror -> injecterror
