+++ pkg/networkservice/utils/null/client.go
